Fix formulate reorg and event workers (#374)
* Test multiple chunks in formulate reorg

* Fix broken StorageLock...

The StorageLocks all have a default deadline, they don't just remain locked.
For some reason (yet to debug), the worker calls are taking a long time / timing out, and due to the broken lock additional worker processes are piling up.

* Dont give up formulating a reorg, even if we are missing local data

* Back to StorageLock with increased deadline

Each WASM worker runs in its own thread and does not have shared memory, so the previous fix only worked for native.

Setting to 120s deadline should work most of the time, and when it doesn't it's not the end of the world to ignore a few irrelevant blocks, just inefficient if it happens too often.
The other side effect is that the storage lock persists across restarts, which is the downside of making it too long.

* Rewrite formulate reorg to process forward/reverse in lockstep

Also fix the handling for traversing before the first block, but err out if the worker is otherwise missing data.

* Bump version