Use EIP-234 getLogs (#386)
This patch switches from using `getLogs` with `fromBlock` and `toBlock` to using `blockHash`, in accordance with EIP-234. This can fix transient issues where we ask for a block, say 5, then we ask for the logs from block #5 but subsequently there's a new block 5 due to re-org and then we can the logs from the wrong block. This may also fix any glitches with Infura transiently sending empty logs when they are not loaded yet.

We also remove an errant console log message.