Add Stack Trace from Call-site not Error site
Due to the async callback nature of sendAndWaitForEvents, errors are raised in the callback code, which leads to an error's stacktrace being just that, in some generic callback code. This patch captures a backtrace at the original callsite and then attaches that to an error before it's returned. As such, it's possible to give the user a more helpful error. It's a bit of a hack, but not utilizing any undefined behavior.